% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mango_utils.R
\name{calculate_cut_site_density}
\alias{calculate_cut_site_density}
\title{Calculate Cut Site Density}
\usage{
calculate_cut_site_density(bins, GRfile)
}
\arguments{
\item{bins}{Bins GRanges object}

\item{GRfile}{Genome GRanges object}
}
\value{
Vector of average scores
}
\description{
This function calculates restriction enzyme cut site density information for given bins.
}
\examples{
\dontrun{
genome_file <- GenomicRanges::GRanges(
  seqnames = "chr1",
  ranges = IRanges::IRanges(start = 1:100, width = 1)
)
bins <- GenomicRanges::GRanges(
  seqnames = "chr1",
  ranges = IRanges::IRanges(start = seq(1, 100, by = 10), width = 10)
)
site_info <- calculate_cut_site_density(bins, genome_file)
}
}
