% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significantLoops.R
\name{convert_df_to_granges}
\alias{convert_df_to_granges}
\title{Convert DataFrame to GRanges Object}
\usage{
convert_df_to_granges(
  df,
  seqnames.field,
  start.field,
  end.field,
  ignore.strand = TRUE
)
}
\arguments{
\item{df}{Data frame to convert}

\item{seqnames.field}{Column name for sequence names}

\item{start.field}{Column name for start positions}

\item{end.field}{Column name for end positions}

\item{ignore.strand}{Whether to ignore strand information}
}
\value{
GRanges object
}
\description{
Convert DataFrame to GRanges Object
}
\examples{
\dontrun{
library(GenomicRanges)
df <- data.frame(
  chr = c("chr1", "chr1", "chr2"),
  start = c(100, 200, 300),
  end = c(150, 250, 350),
  score = c(1, 2, 3)
)
gr <- convert_df_to_granges(df, "chr", "start", "end")
print(gr)
}
}
\keyword{internal}
