% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mango_utils.R
\name{create_equal_bins}
\alias{create_equal_bins}
\title{Create Equal Bins}
\usage{
create_equal_bins(vectortobin, binmethod = "equalsize", numberbins = 30)
}
\arguments{
\item{vectortobin}{Vector to bin}

\item{binmethod}{Method for binning ("equalsize" or "equalocc")}

\item{numberbins}{Number of bins}
}
\value{
Vector of bin borders
}
\description{
This function creates bins for a given vector. This function is directly developed from R package mango.
}
\examples{
\dontrun{
vector_to_bin <- 1:100
bin_borders <- create_equal_bins(vector_to_bin, binmethod = "equalsize", numberbins = 10)
}
}
