% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionHicPro.R
\name{create_interaction_matrix}
\alias{create_interaction_matrix}
\title{Create Interaction Matrix from HiC-pro Output}
\usage{
create_interaction_matrix(intervals, matrix, output)
}
\arguments{
\item{intervals}{Path to the file storing the indices of individual bins of chromosomes.}

\item{matrix}{Path to the file storing the interactions between bins (matrix).}

\item{output}{Path to the output file where the intervals + contact count will be written.}
}
\value{
Invisible NULL
}
\description{
This function creates an interaction matrix (chromosome intervals + contact count)
from a given HiC-pro output matrix.
}
\examples{
\dontrun{
create_interaction_matrix(intervals = "/path/to/intervals.txt",
                          matrix = "/path/to/matrix.txt",
                          output = "/path/to/output_interactions.txt")
}
}
