% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mango_utils.R
\name{create_loops_object_fast}
\alias{create_loops_object_fast}
\title{Create Loops Object Fast}
\usage{
create_loops_object_fast(beddir, snames = NA, type = "all")
}
\arguments{
\item{beddir}{A string. The preprocessed data directory}

\item{snames}{A character vector. Optional list of samples to read in}

\item{type}{Specifies 'intra', 'inter', or 'all' looping. Default 'all'}
}
\value{
A loops object
}
\description{
This function creates loop objects from BEDPE files quickly.
It is an optimized version of the loopsMake function from R package diffloop.
}
\examples{
\dontrun{
beddir <- "/home/user/hichip/bedpe_files"
sname <- "sample1"
snames = paste0(sname, ".filt.intra")
loops <- create_loops_object_fast(beddir, snames = snames, type = "all")
}
}
