% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_local.R
\name{create_washu_local}
\alias{create_washu_local}
\title{Create Local WashU Genome Browser Tracks}
\usage{
create_washu_local(outdir, FDR)
}
\arguments{
\item{outdir}{Character string. Output directory.}

\item{FDR}{Numeric. False Discovery Rate threshold.}
}
\value{
Invisible NULL
}
\description{
This function creates WashU Genome Browser tracks for the local mode of sintHiChIP.
}
\keyword{internal}
