% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_mode_process.R
\name{find_peak_overlaps_global}
\alias{find_peak_overlaps_global}
\title{Find overlaps between HiChIP interactions and peak regions for Global Mode}
\usage{
find_peak_overlaps_global(interactions, peaks_gr)
}
\arguments{
\item{interactions}{Data table with interaction data}

\item{peaks_gr}{GenomicRanges object containing peak regions}
}
\value{
Data table with interactions that overlap with peaks
}
\description{
Find overlaps between HiChIP interactions and peak regions for Global Mode
}
\examples{
\dontrun{
peaks_gr <- GenomicRanges::GRanges(
  seqnames = c("chr1", "chr2"),
  ranges = IRanges::IRanges(start = c(1000, 2000), end = c(1500, 2500))
)
interactions_dt <- data.table(
  chr1 = "chr1", start1 = 1200, end1 = 1300,
  chr2 = "chr1", start2 = 2200, end2 = 2300,
  count = 10
)
overlapped <- find_peak_overlaps_global(interactions_dt, peaks_gr)
}
}
