% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localInteraction.R
\name{generate_hichipper_config}
\alias{generate_hichipper_config}
\title{Generate hichipper Config}
\usage{
generate_hichipper_config(output_file, peaks, resfrags, hicpro_output)
}
\arguments{
\item{output_file}{Character string. Name of the yaml configuration file only.}

\item{peaks}{Character string. Peaks options or path to peak file.}

\item{resfrags}{Character string. Paths to restriction fragments files.}

\item{hicpro_output}{Character string. Paths to HiC-Pro output directories.}
}
\value{
Invisible NULL
}
\description{
This function generates a YAML configuration file for hichipper.
}
\examples{
\dontrun{
generate_hichipper_config(
  output_file = "config.yaml",
  peaks = "EACH,ALL",
  resfrags = "/home/user/hichip/restriction_fragments.bed",
  hicpro_output = "/home/user/hichip/hicpro_output"
)
}
}
