% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_function.R
\name{makecombos}
\alias{makecombos}
\title{Generate Peak Combinations for HiChIP Analysis}
\arguments{
\item{chrom}{Character vector specifying chromosome name}

\item{chrpeaks}{Data frame containing peak information with columns:
\itemize{
  \item chr: Chromosome name
  \item start: Start position
  \item end: End position
  \item score: Peak score
}}

\item{mindist}{Minimum distance between peaks (default: 0)}

\item{maxdist}{Maximum distance between peaks (default: 100000000)}
}
\value{
A data frame containing:
\itemize{
  \item score1: Score of first peak
  \item score2: Score of second peak
  \item dist: Distance between peaks
}
}
\description{
Generate all possible peak combinations within specified distance constraints
}
\details{
Generate Peak Combinations
}
\examples{
peaks <- data.frame(
  chr = rep("chr1", 3),
  start = c(1000, 5000, 10000),
  end = c(2000, 6000, 11000),
  score = c(10, 15, 20)
)
result <- makecombos("chr1", peaks, 1000, 50000)

}
