% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_mode_process.R
\name{matrix_to_bedpe_global}
\alias{matrix_to_bedpe_global}
\title{Convert matrix to bedpe format for sintHiChIP global mode}
\usage{
matrix_to_bedpe_global(
  matrix_file,
  bed_file,
  output_file,
  min_dist = 20000,
  max_dist = 2e+06
)
}
\arguments{
\item{matrix_file}{Path to the matrix file}

\item{bed_file}{Path to the bed file containing bin coordinates}

\item{output_file}{Path to the output bedpe file}

\item{min_dist}{Minimum interaction distance in base pairs (default: 20000)}

\item{max_dist}{Maximum interaction distance in base pairs (default: 2000000)}
}
\value{
Number of interactions written to output file
}
\description{
Convert matrix to bedpe format for sintHiChIP global mode
}
\examples{
\dontrun{
matrix_to_bedpe_global(
  matrix_file = "sample_matrix.matrix",
  bed_file = "sample_matrix_abs.bed", 
  output_file = "sample_temp.bedpe",
  min_dist = 20000,
  max_dist = 2000000
)
}
}
