% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalInteraction.R
\name{overlap_peaks_global}
\alias{overlap_peaks_global}
\title{Process Global (Peak-to-All) Interactions in sintHiChIP Global Mode}
\usage{
overlap_peaks_global(peaks, interaction_file, output_directory, prefix)
}
\arguments{
\item{peaks}{Path to peak file}

\item{interaction_file}{Path to interaction file}

\item{output_directory}{Output directory}

\item{prefix}{Prefix for output files}
}
\value{
Invisible NULL
}
\description{
This function processes the final intra interactions for the downstream significant analysis.
}
\examples{
\dontrun{
overlap_peaks_global(peaks = "/home/user/hichip/peaks.bed",
                     interaction_file = "/home/user/hichip/interactions.txt",
                     output_directory = "/home/user/hichip/output",
                     prefix = "sample1")
}
}
