% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significantLoops.R
\name{sintHiChIP_loops}
\alias{sintHiChIP_loops}
\title{Identify Significant HiChIP Loops}
\usage{
sintHiChIP_loops(
  sname,
  cwd,
  outdir,
  normSiteFile,
  local = TRUE,
  FDR = 0.01,
  nbins = 10
)
}
\arguments{
\item{sname}{Sample name. No default value.}

\item{cwd}{Current working directory. No default value.}

\item{outdir}{Output directory. No default value.}

\item{normSiteFile}{Normalization site file. Should be a BED format file containing normalization cut site density information. No default value.}

\item{local}{Use local method. Default is TRUE.}

\item{FDR}{False Discovery Rate. Default is 0.01.}

\item{nbins}{Number of bins. Default is 10.}
}
\value{
Invisible NULL
}
\description{
This function generates significant loops from HiChIP data.
}
\examples{
\dontrun{
sintHiChIP_loops(sname = "sample1",
                 cwd = "/home/user/hichip/working/directory",
                 outdir = "/home/user/hichip/output",
                 normSiteFile = "/home/user/hichip/normalization/file.bed",
                 local = TRUE,
                 FDR = 0.01,
                 nbins = 10)
}
}
